requirements.txt
fastapi
uvicorn
gunicorn
psycopg2
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
